Simplify Column and Codec
Turn Column and Codec from traits into simple structs.  Column is now
composed mainly of a list of data sections, and a Codec.  Codec does not
any data anymore, and is a description of the decoding scheme composed
from a number of predefined encoding operations.

This change also allows elimination of the lifetime parameters on Codec,
Type and QueryPlan. Additionaly, it was necessary to generalize the
query planner to allow individual operations to have both streaming +
nonstreaming inputs and automatically perform a topological sort on
query stages to determine execution order.